pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel", "setuptools-scm>=9"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
asthralios = "lib/asthralios"

[tool.setuptools.packages.find]
where = ["lib"]

[project]
name = "asthralios"
dynamic = ["version"]
description = "Markizano's Assistant"
readme = "README.md"
requires-python = ">=3.11, <=3.13"
license = "GPL-3.0-or-later"
authors = [
    { name = "Markizano Draconus", email = "python@markizano.net" },
]
maintainers = [
    { name = "Markizano Draconus", email = "python@markizano.net" },
]
dependencies = [
    "atlassian-python-api==4.0.7",
    "discord-py==2.7.1",
    "easydict>=1.13",
    "faster-whisper>=1.2.1",
    "kizano==1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-ollama>=1.0.1",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=0.3.1",
    "langchain>=1.2.13",
    "langgraph>=1.0.8",
    "openai>=2.21.0",
    "openwebui-client==0.3.2",
    "pasimple==0.0.3",
    "pygithub==2.7.0",
    "python-dotenv==1.2.2",
    "pyyaml>=6.0.1",
    "requests>=2.32.5",
    "slack-bolt==1.27.0",
    "slack-sdk==3.41.0",
    "torch>=2.11.0",
    "vobject>=0.9.9",
]
keywords = [ "ai", "artificial intelligence", "agentic", "langchain" ]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:'audioop' is deprecated and slated for removal in Python 3.13:DeprecationWarning",
]

[project.urls]
homepage = "https://markizano.net/"
repository = "https://github.com/markizano/asthralios"

[project.scripts]
asthralios = "asthralios:main"
